model_api/requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.16.1
numpy
pillow